Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -4,7 +4,7 @@ int main() {
     int a,s,d;
     float av;
     scanf("%d %d %d",&a,&s,&d);
-    av = (a + s + d)/3;
+    av = (float)(a + s + d)/3;
     printf("Average: %.2f",av);
     return 0;
 }
